Port check_enter_force_ums_mode from debian u-boot
For Tinker Board S, this commit implements a function that
the device can enter UMS mode when it is connected to PC via USB cable.
This is from https://github.com/TinkerBoard/debian_u-boot .
As there is no check for device type in source code, it won't work with
other devices.